+++ apps/worxpace/src/constants/site.ts
@@ -4,14 +4,3 @@ export const siteConfig = {
 };
 
 export const MAX_SHOWN_USERS = 2;
-
-export const GROUPS = [
-  "document",
-  "kanban",
-  "whiteboard",
-  "workflow",
-  "trash:document",
-  "trash:kanban",
-  "trash:whiteboard",
-  "trash:workflow",
-];