+++ apps/worxpace/src/hooks/use-documents.ts
@@ -5,7 +5,8 @@ import { toast } from "sonner";
 import useSWR from "swr";
 import useSWRMutation from "swr/mutation";
 
-import type { Page } from "@swy/notion";
+import type { DocItemData } from "@swy/notion";
+import { useTree } from "@swy/ui/shared";
 
 import {
   archiveDocument,
@@ -14,7 +15,12 @@ import {
   restoreDocument,
   updateDocument,
 } from "~/actions";
-import { documentsFetcher, toPage, type DetailedDocument } from "~/lib";
+import {
+  documentsFetcher,
+  toDocItem,
+  toPage,
+  type DetailedDocument,
+} from "~/lib";
 
 export const useDocuments = ({
   workspaceId,
@@ -22,6 +28,7 @@ export const useDocuments = ({
   workspaceId?: string | null;
 }) => {
   const router = useRouter();
+  const setNodes = useTree((state) => state.set<DocItemData>);
   /** Fetcher */
   const key = workspaceId ? { type: "document" as const, workspaceId } : null;
   const {
@@ -30,12 +37,12 @@ export const useDocuments = ({
     error,
     mutate,
   } = useSWR<DetailedDocument[], Error>(key, documentsFetcher, {
+    onSuccess: (data) => setNodes(data.map(toDocItem)),
     onError: (e) => console.log(`[swr:document]: ${e.message}`),
   });
-  const fetchPages = async (): Promise<Page[]> => {
+  const fetchPages = async () => {
     const documents = await mutate();
-    if (!documents) return [];
-    return documents.map((document) => toPage(document)!);
+    return documents?.map((document) => toPage(document)!) ?? [];
   };
   /** Mutations */
   const onError = (e: Error) => toast.error(e.message);