+++ packages/notion/src/sidebar/index.ts
@@ -1,2 +1,3 @@
 export * from "./modals";
 export * from "./sidebar";
+export type * from "./types";