+++ packages/ui/src/components/shared/tree/index.tsx
@@ -1,5 +1,0 @@
-export * from "./index.types";
-export * from "./tree-actions";
-export * from "./tree-context";
-export * from "./tree-list";
-export * from "./tree-provider";